+++ src/components/Restaurant/Restaurant.jsx
@@ -1,24 +1,14 @@
-import React, { useState } from "react";
+import React from "react";
 import "./Restaurant.scss";
-import RestaurantButton from "../RestaurantButton/RestaurantButton";
+import Order from "../Order/Order";
 
 const Restaurant = () => {
-  const [orders, setOrders] = useState(0);
-
-  const orderOne = () => {
-    setOrders(orders + 1);
-  };
-
   return (
     <section className="restaurant">
       <h3 className="restaurant__heading">Restaurant Orders</h3>
       <ul className="restaurant__list">
-        <li className="restaurant__item">
-          Pizzas: {orders}{" "}
-          <RestaurantButton orderOne={orderOne} />
-        </li>
+        <Order />
       </ul>
-
     </section>
   );
 };
